Exploratory Data Analysis with Python and Pandas
Exploratory Data Analysis With Python and Pandas. This is a project-based course which should take approximately 2 hours to finish. Before diving into the project, please take a look at the course objectives and structure:
#Course Objectives
In this course, we are going to focus on three learning objectives:
1.	Apply practical Exploratory Data Analysis (EDA) techniques on any tabular dataset using Python.
2.	Produce data visualizations using Seaborn and Matplotlib
3.	Identify and handle duplicate and missing data
By the end of the course, you will be able to dive in and explore a sales dataset from a supermarket retail chain, and apply those skills to any tabular dataset using Python. You will learn to use external Python packages such as Pandas, Numpy, Seaborn, Matplotlib and more to retrieve valuable information from your dataset.